+++ packages/core/CHANGELOG.md
@@ -4,6 +4,14 @@
 
 - Remove BN.js in favor of native BigInt.
 
+## 1.16.1 (2022-06-16)
+
+- Fix VM execution error in `proposeUpgrade` with Gnosis Chain. ([#597](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/597))
+
+## 1.16.0 (2022-06-16)
+
+- Return ethers transaction response with `proposeUpgrade`. ([#554](https://github.com/OpenZeppelin/openzeppelin-upgrades/issues/554))
+
 ## 1.15.0 (2022-05-31)
 
 - Support `unsafeSkipStorageCheck` option in `ValidationOptions`. ([#566](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/566))